Server seongwol (#30)
* [Config | constructor] | fix: #처리가 급하게 필요해서 index[0] 이 # 일 때 제끼도록 해서 서버블록 두개를 잡는 문제 해결

* [project] refactor : 코드 리펙토링 및 플래그 적용

* [project] | feat: makefile 추가

+++ Parse/Config.hpp
@@ -21,7 +21,7 @@ public:
 
     ~Config();
     // void setServers(const std::map<std::string, ServerConfig> servers);
-    const ServerConfig& operator[](int index) const;
+    const ServerConfig& operator[](size_t index) const;
     void    parseConfig(const std::string filename);
     int getNumberOfServer() const;
 };
